% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.summary.R
\name{bootstrap.summary}
\alias{bootstrap.summary}
\title{Based on the output obtained from ipw.bootstrap.ph2 function, empirical distribution based-variance/-confidence intervals for the estimated regression coefficients,
prevalences, cumulative sub-distribution hazards and cumulative incidences for events 1 and 2 are calculated.}
\usage{
bootstrap.summary(o.input, b.input, p.mat, i.mat1, i.mat2, time.points,
  alpha = 0.05, ...)
}
\arguments{
\item{o.input}{the model fit output obtained from ipw.pi.competing function}

\item{b.input}{the bootstrap output obtained from ipw.bootstrap.ph2 function}

\item{p.mat}{design matrix for predicting prevalence by using the inverse of logit function; 
both of vector and matrix types are allowed; the first component or column should include 1 for the intercept.}

\item{i.mat1}{design matrix for predicting cumulative sub-distribution hazards and cumulative incidences for event 1}

\item{i.mat2}{design matrix for predicting cumulative sub-distribution hazards and cumulative incidences for event 2}

\item{time.points}{time points at which cumulative sub-distribution hazards and cumulative incidences for events 1 and 2 are predicted}

\item{alpha}{The nominal coverage probability is (1-alpha)*100\%. Default to 0.05}
}
\value{
The output is a list of class ipw.bootstrap.ph2, which contains the following elements.
 \itemize{
 \item reg.covariance  empirical distribution based-covariance for the regression coefficients included in o.input
 \item reg.coef.ci empirical distribution based-confidence intervals for the regression coefficients included in o.input
 \item prev.ci empirical distribution based-confidence intervals for the predicted prevalence
 \item subdist.hazard1.ci confidence intervals for the predicted cumulative sub-distribution hazard function for event 1
 \item subdist.hazard2.ci confidence intervals for the predicted cumulative sub-distribution hazard function for event 2
 \item cum.inc1.ci confidence intervals for the predicted cumulative incidences for event 1
 \item cum.inc2.ci confidence intervals for the predicted cumulative incidences for event 2
 \item trans.r the transformation parameters used in the bootstrap summary
 \item p.mat the input information, the design matrix for prevalence
 \item i.mat1 the input information, the design matrix for cumulative sub-distribution hazard and cumulative incidence for event 1
 \item i.mat2 the input information, the design matrix for cumulative sub-distribution hazard and cumulative incidence for event 1
 }
}
\description{
Based on the output obtained from ipw.bootstrap.ph2 function, empirical distribution based-variance/-confidence intervals for the estimated regression coefficients,
prevalences, cumulative sub-distribution hazards and cumulative incidences for events 1 and 2 are calculated.
}
\references{
\itemize{
 \item Hyun N, Katki HA, Graubard BI.
       Sample-Weighted Semiparametric Estimation of Cause-Specific Cumulative Risk and Incidence Using Left or Interval-Censored Data from Electronic Health Records. Statistics in Medicine 2020;
       under the 2nd review.
}
}
\author{
Noorie Hyun, \email{nhyun@mcw.edu}, Xiao Li \email{xiaoli@mcw.edu}
}
