% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.bootstrap.ph2.R
\name{ipw.bootstrap.ph2}
\alias{ipw.bootstrap.ph2}
\title{To generate the empirical distribution for estimated sub-distribution hazard functions and regression coefficients, 
we adopted the two-phase weighted bootstrap procedure (Hyun et al, reference). We fit ipw.pi.competing function from main.R to 
the bootstrap weighted data.The most arguments except for three arguments, n.boot, parallel, n.core are necessary for fitting "ipw.pi.competing" function.}
\usage{
ipw.bootstrap.ph2(n.boot, Data, p.model, i.model1, i.model2,
  trans.r1 = 0, trans.r2 = 0, n.beta = 1, n.gamma1 = 0,
  n.gamma2 = 0, reg.initials = NULL, convergence.criteria = 0.001,
  iteration.limit = 250, time.interval, time.list = NULL,
  parallel = FALSE, n.core = NULL, ...)
}
\arguments{
\item{n.boot}{the number of replications for bootstrap resampling}

\item{parallel}{if parallel=TRUE, n.core should be specified. Default to FALSE.}

\item{n.core}{the nummber of cores for parallel computing}
}
\value{
The output is a list of class ipw.bootstrap.ph2, which contains the following elements.
 \itemize{
 \item reg.coef: estimated regression coefficients
 \item time finite time points at which subdistribution hazards are estimated
 \item subdist.hazard1 estimated subdistribution hazard function for event 1
 \item subdist.hazard2 estimated subdistribution hazard function for event 2
 }
}
\description{
To generate the empirical distribution for estimated sub-distribution hazard functions and regression coefficients, 
we adopted the two-phase weighted bootstrap procedure (Hyun et al, reference). We fit ipw.pi.competing function from main.R to 
the bootstrap weighted data.The most arguments except for three arguments, n.boot, parallel, n.core are necessary for fitting "ipw.pi.competing" function.
}
\references{
\itemize{
 \item Hyun N, Katki HA, Graubard BI.
       Sample-Weighted Semiparametric Estimation of Cause-Specific Cumulative Risk and Incidence Using Left or Interval-Censored Data from Electronic Health Records. Statistics in Medicine 2020;
       under the 2nd review.
}
}
\author{
Noorie Hyun, \email{nhyun@mcw.edu}, Xiao Li \email{xiaoli@mcw.edu}
}
