% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ipw.pi.competing}
\alias{ipw.pi.competing}
\title{Sample-Weighted Prevalence-Incidence Mixture Models for Competing Incident Events (as of 03/03/2020)}
\usage{
ipw.pi.competing(Data, p.model, i.model1, i.model2, trans.r1 = 0,
  trans.r2 = 0, n.beta = 1, n.gamma1 = 0, n.gamma2 = 0,
  reg.initials = NULL, convergence.criteria = 0.001,
  iteration.limit = 250, time.interval = 0.1, time.list = NULL,
  population = "super", anal.var = TRUE, ...)
}
\arguments{
\item{Data}{Data used to fit the model containing columns for each term in p.model, i.model1 and i.model2 expressions.
For stratified random sampling designs, columns denoted samp.wgt and strata are expected indicating the sampling weights and sampling strata.
population="super" option, an additional column denoted strata.frac is expected indicating the fraction of the population
that consists of each strata.  For example, if in the target population there are three strata that occurs with proportions 0.2, 0.4, and 0.6,
then strata.frac will take values of 0.2, 0.4 or 0.6.}

\item{p.model}{The prevalence model for event 1 to be fitted, specified using an expression of the form \emph{C~model}.
Elements in the expression are as followed:
\itemize{
\item c - Numeric variable indicating whether the event was prevalent at time zero, taking values of 1="Yes", 0="No", -999="Unknown";
\item model - Linear predictor consisting of a series of terms separated by \emph{+} operators.
}}

\item{i.model1}{The incidence model for event 1 to be fitted using an expression of the form \emph{C+L1+R1~model1}
\itemize{
\item C - Numeric variable indicating whether the event was prevalent at time zero, taking values of 1="Yes", 0="No", -999="Unknown";
\item L1 - Numeric starting time of the interval in which event 1 occurred, with -999 denoting known prevalent events;
\item R1 - Ending time of the interval in which event 1 occurred, with -999 and Inf denoting known prevalent event 1 and right-censoring, respectively;
\item model1 - Linear predictor consisting of a series of terms separated by \emph{+} operators.
}}

\item{i.model2}{The incidence model for event 2 to be fitted, specified using an expression of the form \emph{L2+R2~model2}
\itemize{
 \item L2 - Numeric starting time of the interval in which event 1 occurred, with -999 denoting known prevalent event 1;
 \item R2 - Ending time of the interval in which event 1 occurred, with -999 and Inf denoting known prevalent event 1 and right-censoring, respectively;
 \item model2 - Linear predictor consisting of a series of terms separated by \emph{+} operators.
 }}

\item{trans.r1}{The parameter "r" for the transformation function for event 1, G(x)=log(1+rx)/r for r>0;G(x)=x for r=0 (default),which indicates proportional hazards model for the subdistribution hazard function.}

\item{trans.r2}{The parameter "r" for the transformation function for event 2. Default to 0.}

\item{n.beta}{is The number of regressors expressed in the p.model plus 1 (for intercept). If p.model is "C~1", n.beta=1.}

\item{n.gamma1}{The number of regressors expressed in the i.model1. If i.model1 is "C+L1+R1~1", n.gamma1=0.}

\item{n.gamma2}{The number of regressors expressed in the i.model2. If i.model2 is "L2+R2~1", n.gamma2=0.}

\item{reg.initials}{The initial values for regression coefficients in the order of (p.model, i.model1, i.model2). The number of components for reg.initials is n.beta+n.gamma1+n.gamma2. Default to be NULL.}

\item{convergence.criteria}{The criterion for the convergence of the iterated algorithm. Default to 0.001}

\item{iteration.limit}{The maximum number allowed for the iteration of the algorithm. Default to 250.}

\item{time.interval}{time.interval determines how finner finite time points are evenly divided, at which subdistribution hazard functions are estimated. Default to 0.1.}

\item{time.list}{a vector of finite time points at which subdistribution hazard functions are estimated. Default to NULL. For example, when an irregular spaced time points are of interest, time.list=c(1,3,8,10).}

\item{population}{options="super" and "finite" include variation due to super-population sampling and finite sampling from the super-population and variation due to finite sampling from a finite population, respectively. Default to "super".}

\item{anal.var}{analytical variance estimation is provided when anal.var=TRUE and the inverse information matrix exists. Default to TRUE.}
}
\value{
The output is a list of class ipw.pi.competing.risks, which contains the following elements.
 \itemize{
 \item data.summary: A data frame containing the following:
   Included subjects - number of observations with complete data;
   Known prevalent event 1 - the number of events known to be prevalent at time zero;
   Incident event 1 - the number of event times for event 1 occuring in the interval (L1>=0,R1<Inf] and C=0;
   Incident event 2 - the number of event times for event 2 occuring in the interval (L2>=0,R2<Inf] and C=0;
   Left censored event 1 - the number of event times known to occur by R1<Inf, but can also have been prevalent at time zero, that is C=-999;
   Right censoring - the number of observations right-censored with event time occurring in the interval (L1>0,R1=Inf) or (L2>0,R2=Inf) with C=0;
   Missing prevalent+right censoring - the number of observations with intervals (0,Inf) for both events 1 and 2 and C=-999.
   Maximum follow-up time for event 1
   Maximum follow-up time for event 2
   Maximum right censoring time for event 1
   Maximum right censoring time for event 2
 \item reg.coef: A data frame summarizing parameter values, standard errors, and 95 percent confidence intervals.
 \item reg.covariance: The analytical asymptotic covariance matrix exists for the regression coefficient estimates
 \item prevalence: A vector of prevalence estimates given covariates specified in p.model 
 \item subdist.hazard1: A data frame includes two columns, time and estimated subdistribution hazard for event 1.
 \item subdist.hazard2: A data frame includes two columns, time and estimated subdistribution hazard for event 2.
 \item subdist.hazard.fn1: A function returns estimated subdistribution hazard for event 1 when time points are input.
 \item subdist.hazard.fn2: A function returns estimated subdistribution hazard for event 2 when time points are input.
 \item convergence: Convergence statistics
 \item run.time.mins: The elapsed computation time in minutes.
 \item loglike: Sample-weighted log-likelikelihood at the estimated parameters
 \item trans.r: A vector of the specified parameters for the transformation functions used for events 1 and 2
 \item models: A vector of the specified models, p.model, i.model1 and i.model2
 }
}
\description{
This package fits competing risks models to failure time (or survival time) data for two competing events. 
Failure time for one event of the competing events can be prevalent left-censored, interval-censored or a mixture 
of truly incident disease and missed prevalent disease when disease ascertainment is not always conducted at baseline, 
while failure time for the other event is only interval-censored. Baseline is set to be time 0. 
General transformation,G(x)=(1+r*x)/r if r>0; =x if r=0, is used for a subdistribution hazard function multiplied by an exponential effect of a linear combination of risk factors  
for flexible incidence models. Logistic regression models are used for prevalence. The IPW log-likelihood approach, which uses the inverse of sample inclusion probabilities,
is employed to account for different sampling fractions across strata.
}
\references{
\itemize{
 \item Hyun N, Katki HA, Graubard BI.
       Sample-Weighted Semiparametric Estimation of Cause-Specific Cumulative Risk and Incidence Using Left or Interval-Censored Data from Electronic Health Records. Statistics in Medicine 2020;
       under the 2nd review.
}
}
\author{
Noorie Hyun, \email{nhyun@mcw.edu}, Xiao Li \email{xiaoli@mcw.edu}
}
