% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.predict.R
\name{pi.pred}
\alias{pi.pred}
\title{Based on the output obtained by applying ipw.pi.competing function to data, 
pi.predict function calculates predicted prevalence by using the inverse function of logit function, 
cumulative sub-distribution hazards and cumulative incidences for events 1 and 2.}
\usage{
pi.pred(input, p.mat, i.mat1, i.mat2, time.points)
}
\arguments{
\item{input}{the output obtained from ipw.pi.competing}

\item{p.mat}{design matrix for predicting prevalence by using the inverse of logit function; 
both of vector and matrix types are allowed; the first component or column should include 1 for the intercept.}

\item{i.mat1}{design matrix for predicting cumulative sub-distribution hazards and cumulative incidences for event 1}

\item{i.mat2}{design matrix for predicting cumulative sub-distribution hazards and cumulative incidences for event 2}

\item{time.points}{time points at which cumulative sub-distribution hazards and cumulative incidences for events 1 and 2 are predicted}
}
\value{
The output is a list of class pi.predict, which contains the following elements.
 \itemize{
 \item prev predicted prevalence for the sub-groups with the covariates defined in p.mat
 \item subdist.hazard1 predicted cumulative sub-distribution hazards for event 1 for the sub-group with the covariates defined in i.mat1
 \item subdist.hazard2 predicted cumulative sub-distribution hazards for event 2 for the sub-group with the covariates defined in i.mat2
 \item cum.inc1 predicted cumulative incidences for event 1 for the sub-group with the covariates defined in i.mat1
 \item cum.inc2 predicted cumulative incidences for event 2 for the sub-group with the covariates defined in i.mat2
 }
}
\description{
Based on the output obtained by applying ipw.pi.competing function to data, 
pi.predict function calculates predicted prevalence by using the inverse function of logit function, 
cumulative sub-distribution hazards and cumulative incidences for events 1 and 2.
}
\references{
\itemize{
 \item Hyun N, Katki HA, Graubard BI.
       Sample-Weighted Semiparametric Estimation of Cause-Specific Cumulative Risk and Incidence Using Left or Interval-Censored Data from Electronic Health Records. Statistics in Medicine 2020;
       under the 2nd review.
}
}
\author{
Noorie Hyun, \email{nhyun@mcw.edu}, Xiao Li \email{xiaoli@mcw.edu}
}
