% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samp.data}
\alias{samp.data}
\title{Example data for demo}
\format{a data frame with columns:
\describe{
  \item{C}{Prevalence indicator: 1 = prevalence; 0 = no prevalence}
  \item{L}{Left time point}
  \item{R}{}
  \item{HR_status}{}
  \item{risk_score}{}
  \item{event}{}
  \item{L1}{}
  \item{R1}{}
  \item{L2}{}
  \item{R2}{}
  \item{strata}{}
  \item{samp.wgt}{}
  \item{s.frc}{}
  \item{strata.frac}{}
}}
\source{
simulation
}
\usage{
samp.data
}
\description{
Example data for demo
}
\examples{
\dontrun{
 samp.data
}
}
\keyword{datasets}
